+++ examples/example-antd-management-fast-framework/src/components/TestComponent/index.jsx
@@ -23,6 +23,4 @@ class TestComponent extends PureComponent {
   }
 }
 
-TestComponent.defaultProps = {};
-
 export { TestComponent };