+++ packages/antd-management-fast-component/src/components/Canvas/Spirit/index.jsx
@@ -70,6 +70,4 @@ class Spirit extends Core {
   };
 }
 
-Spirit.defaultProps = {};
-
 export { Spirit };