+++ packages/antd-management-fast-component/src/components/PageExtra/CardCollectionItemContent/index.jsx
@@ -236,7 +236,7 @@ class CardCollectionItemContent extends PureComponent {
                     <ElasticityTree
                       onSelect={onSelectCallback}
                       innerProps={otherTreeSelectProperties}
-                      listData={treeSelectListData}
+                      listData={[...treeSelectListData]}
                       dataConvert={treeSelectDataConvertor}
                     />
                   </Col>