+++ packages/antd-management-fast-design-react/src/components/designable/react/simulators/MobileSimulator/body.jsx
@@ -75,5 +75,3 @@ export const MobileBody = observer((properties) => {
     </div>
   );
 });
-
-MobileBody.defaultProps = {};