+++ packages/antd-management-fast-flow/src/components/Flow/BackwardEdge/index.jsx
@@ -158,6 +158,4 @@ const BackwardEdge = ({
   );
 };
 
-BackwardEdge.defaultProps = {};
-
 export { BackwardEdge };