+++ packages/antd-management-fast-flow/src/components/Flow/CarbonCopyEdge/index.jsx
@@ -140,6 +140,4 @@ class CarbonCopyEdge extends PureComponent {
   }
 }
 
-CarbonCopyEdge.defaultProps = {};
-
 export { CarbonCopyEdge };