+++ packages/antd-management-fast-flow/src/components/Flow/EndNode/index.jsx
@@ -150,6 +150,4 @@ const EndNode = (properties) => {
   );
 };
 
-EndNode.defaultProps = {};
-
 export { EndNode };