+++ packages/antd-management-fast-flow/src/components/Flow/ForwardEdge/index.jsx
@@ -158,6 +158,4 @@ class ForwardEdge extends PureComponent {
   }
 }
 
-ForwardEdge.defaultProps = {};
-
 export { ForwardEdge };