+++ packages/antd-management-fast-flow/src/components/Flow/StartNode/index.jsx
@@ -127,6 +127,4 @@ const StartNode = (properties) => {
   );
 };
 
-StartNode.defaultProps = {};
-
 export { StartNode };