+++ packages/antd-management-fast-framework/src/components/MobileContainor/MobilePreviewDrawer/index.jsx
@@ -156,6 +156,4 @@ class MobilePreviewDrawer extends BaseNeedlessLoadDrawer {
   };
 }
 
-MobilePreviewDrawer.defaultProps = {};
-
 export { MobilePreviewDrawer };