+++ packages/antd-management-fast-framework/src/framework/DataDrawer/Base/index.jsx
@@ -617,8 +617,4 @@ class Base extends BaseWindow {
   }
 }
 
-Base.defaultProps = {
-  flag: '',
-};
-
 export { Base };