+++ packages/antd-management-fast-framework/src/framework/DataDrawer/BaseFormDrawer/index.jsx
@@ -93,6 +93,4 @@ class BaseFormDrawer extends Base {
   };
 }
 
-BaseFormDrawer.defaultProps = {};
-
 export { BaseFormDrawer };