+++ packages/antd-management-fast-framework/src/framework/DataDrawer/BaseVerticalFlexDrawer/index.jsx
@@ -159,6 +159,4 @@ class BaseVerticalFlexDrawer extends Base {
   };
 }
 
-BaseVerticalFlexDrawer.defaultProps = {};
-
 export { BaseVerticalFlexDrawer };