+++ packages/antd-management-fast-framework/src/framework/DataListView/EmptyCardCollection/index.jsx
@@ -18,6 +18,4 @@ class EmptyCardCollection extends PureComponent {
   }
 }
 
-EmptyCardCollection.defaultProps = {};
-
 export { EmptyCardCollection };