+++ packages/antd-management-fast-flow/CHANGELOG.md
@@ -1,5 +1,13 @@
 # antd-management-fast-flow
 
+## 1.0.79
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-component@2.6.177
+
 ## 1.0.78
 
 ### Patch Changes